Use libxkbcommon to apply keyboard mappings
On my machine, Ctrl and Caps are swapped through the systems setting
dialog in gnome or kde. The key mappings are applied after they are read
from the kernel, which is where we are reading from

We need to apply the same mappings manually

* Pull in libxkbcommon
* For now, pass in the keymap manually. This can be generated with
  xkbcomp
* Replace the manual mapping of keycodes to characters with the xkb
  mappings

This also improves the UI of which characters are rendered. Caps
modifiers are applied to characters, and there are more/better names