Automatically load xkb mappings
libxkbcomp is integrated into winit/wayland-client which is pretty far
down the stack. We don't have access to the real mappings, and we need
them to apply user mappings to the evdev codes pulled in from the
kernel. Use xkbcomp to export the mappings through the X/XWayland and
re-load them manually